# app/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is TODO API by golang.",
        "title": "TODO API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/users": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "全てのユーザーのID・名前をリストで取得する",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "全てのユーザーを取得する",
                "responses": {
                    "200": {
                        "description": "登録されたユーザーの情報",
                        "schema": {
                            "$ref": "#/definitions/presenter.SuccessResponse-array_user_GetUsersResponse"
                        }
                    },
                    "400": {
                        "description": "不正なリクエスト",
                        "schema": {
                            "$ref": "#/definitions/presenter.FailureResponse"
                        }
                    },
                    "500": {
                        "description": "内部サーバーエラー",
                        "schema": {
                            "$ref": "#/definitions/presenter.FailureResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "新しいユーザーを登録する",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "ユーザーの登録",
                "parameters": [
                    {
                        "description": "ユーザー登録のための情報",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.PostUserRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "登録されたユーザーの情報",
                        "schema": {
                            "$ref": "#/definitions/presenter.SuccessResponse-user_PostUserResponse"
                        }
                    },
                    "400": {
                        "description": "不正なリクエスト",
                        "schema": {
                            "$ref": "#/definitions/presenter.FailureResponse"
                        }
                    },
                    "500": {
                        "description": "内部サーバーエラー",
                        "schema": {
                            "$ref": "#/definitions/presenter.FailureResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "presenter.FailureResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "presenter.SuccessResponse-array_user_GetUsersResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/user.GetUsersResponse"
                    }
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "presenter.SuccessResponse-user_PostUserResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/user.PostUserResponse"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "user.GetUsersResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "user.PostUserRequest": {
            "type": "object",
            "required": [
                "email",
                "name",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "user.PostUserResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}